tests: ensure that macro example compiles test function
Before the #[test] attribute would be attached to the test_readme_deps
and since code examples in doctests are not compiles with the --test
flag, the test_readme_deps function ended up being ignored!

We now use a hidden function as the "target" for #[test]. The test
function (test_readme_deps) is thus a normal function and will
actually be compiled as part of the doctest.