version_matches_request: consider pre-release identifiers
Before we completely ignored the pre-release identifiers when
comparing versions. This meant that a TOML block with

  [dependencies]
  foo = 1.0.0

was considered to match a crate version of 1.0.0-rc1. That was not
intended: per the semver specification, a pre-release version comes
before the normal version. So depending on version 1.0.0 would not
actually make it possible to install version 1.0.0-rc1.

Fixes #19.