check_html_root_url: attempt to find html_root_url in source
The syn crate does not preserve span information while parsing the
Rust code, so we no longer have access to accurate information about
where each attribute begins and ends.

This adds a simple heuristic: we simply search for html_root_url in
the source code and use that line in the error message.