codecov: allow a drop of up to 5% in overall coverage
By default, Codecov will fail a PR if the project coverage drops even
slightly. This now allows it to change a little.

The change is necessary before moving from Travis CI to CircleCI since
the Docker image used on CircleCI is slightly different from what we
get when we install cargo-tarpaulin on Travis CI.