check_contains_regex: escape placeholders in template
Before, the pkg_version and pkg_name was used directly, which meant
that characters such as "." would retain their special meaning in the
resulting regex. This in turn meant that searching for

  ^Version {version}

would also match a line like

  Version 1.203

when the version number of the current package is 1.2.3, since the "."
in the version number accidentally matches the "0".

Since we will now see more backslashes in the output, we no longer use
Rust escaping when printing the regular expressions to the user.