Drop Rust 2018 compatibility
I've been trying to keep the crate compatible with Rust 2018, which
corresponds to Rust version 1.31.0. However, this has proven
impossible because of how our dependencies keep releasing new patch
versions that push up the minimum required Rust version.

Because of this, the current code no longer compiles with Rust 1.31.0
because `cargo sync` will pull in too new versions of the direct and
transitive dependencies. This happens even if there are no changes in
`version-sync`.

The failure to compile shows up as spurious and unexpected build
failures in our CI, which in turn makes it hard to trust CI.

I'm concluding that it's infeasible to keep `version-sync` compatible
with any particular version of Rust. While not perfect, we'll track
the latest stable version of Rust from now on. This is not perfect
since even stable Rust sometimes decides to turn warnings into hard
errors: https://github.com/rust-lang/rust/pull/64221/