Parse TOML blocks inside blockquotes
If a TOML block was found inside a Markdown blockquote, we would
report a TOML parse error. This is because we extracted the code block
directly from the original text and thus incorrectly included the `>`
characters.

So parsing

~~~
> This is a blockquote
>
> ```toml
> clap = { git = "https://github.com/clap-rs/clap/" }
> ```
~~~

would result in an error like this:

> README.md (line 22) ... TOML parse error: unexpected character
>  found: `>` at line 1 column 1 in
>
> `> clap = { git = "https://github.com/clap-rs/clap/" }`

To fix this, we now parse the code blocks by listening to the `Start`
and `End` events from `pulldown-cmark`. This should be much more
robust.

Fixes #83.