appveyor: disable caching
It takes about 60 seconds to restore the cache at the beginning of the
build and about 90 seconds to store it afterwards. In contrast, it
takes about 70 seconds to download and compile all dependencies from
scratch.

Even though the cache is only updated when building from the master
branch, the savings from the cache are at best minimal.