Check version numbers in `=1.2.3` and `>1.2.3`, and `>=1.2.3`
Before, a dependency specified as

```toml
[dependencies]
foo = ">= 1.2.3, < 2.0"
```

was ignored since it was deemed too complicated to check versions
requirements with multiple comparators.

We now handle such cases by checking every comparator individually.
Comparators which form a lower-bound for the version is checked. This
means that version requirements of the form

* `1.2` and `^1.2`
* `~1.2`
* `=1.2`
* `>1.2`
* `>=1.2`
* `1.2.*`

are all checked for compatibility with the actual crate version. If
the crate version is higher than the version requirement, an error is
returned.

Upper-build version requirements such as `<2.0` are ignored like
before.